Got the new search to be as good or better than the old one

+++ src/main/java/org/gainratio/amlfilter/model/Result.java
@@ -9,6 +9,6 @@ public class Result {
     private String resultName;
     private String entityCodeInSource;
     private String listName;
-    private Float textSimilarity;
-    private Float resultNameInformationLevel;
+    private Double textSimilarity;
+    private Double resultNameInformationLevel;
 }
